+++ src/app/event/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2025 Gabriel Bjørnager Jensen.
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum Event {
+	Terminate,
+}