+++ src/app/mod.rs
@@ -2,6 +2,10 @@
 
 mod app;
 mod config;
+mod event;
+mod graphics_context;
 
 pub use app::App;
 pub use config::Config;
+pub use event::Event;
+pub use graphics_context::GraphicsContext;