+++ src/level/block/mod.rs
@@ -1,5 +1,9 @@
 // Copyright 2025 Gabriel Bjørnager Jensen.
 
+use crate::level::BlockFromStrError;
+
+use std::str::FromStr;
+
 #[repr(u8)]
 #[derive(Clone, Copy, Debug, Default, Eq, PartialEq)]
 pub enum Block {
@@ -19,4 +23,30 @@ pub enum Block {
 	Marble,
 	Limestone,
 	Grass,
+}
+
+impl FromStr for Block {
+	type Err = BlockFromStrError;
+
+	#[inline]
+	fn from_str(s: &str) -> Result<Self, Self::Err> {
+		match s {
+			"air"       => Ok(Self::Air),
+			"basalt"    => Ok(Self::Basalt),
+			"bedrock"   => Ok(Self::Bedrock),
+			"clay"      => Ok(Self::Clay),
+			"dirt"      => Ok(Self::Dirt),
+			"granite"   => Ok(Self::Granite),
+			"grass"     => Ok(Self::Grass),
+			"gravel"    => Ok(Self::Gravel),
+			"limestone" => Ok(Self::Limestone),
+			"magma"     => Ok(Self::Magma),
+			"marble"    => Ok(Self::Marble),
+			"sand"      => Ok(Self::Sand),
+			"stone"     => Ok(Self::Stone),
+			"water"     => Ok(Self::Water),
+
+			_ => Err(BlockFromStrError { name: s.into() })
+		}
+	}
 }