+++ src/level/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2025 Gabriel Bjørnager Jensen.
 
 mod block;
+mod block_from_str_error;
 mod chunk;
 mod columns;
 mod columns_mut;
@@ -9,6 +10,7 @@ mod map;
 mod map_size;
 
 pub use block::Block;
+pub use block_from_str_error::BlockFromStrError;
 pub use chunk::Chunk;
 pub use columns::Columns;
 pub use columns_mut::ColumnsMut;