+++ src/version/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2025 Gabriel Bjørnager Jensen.
+
+mod version;
+
+pub use version::Version;