+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This is the changelog of Bedrock.
 See `README.md` for more information.
 
+## 0.3.0
+
+* Finalise custom level loading
+* Add basic graphics
+* Set terminate handler
+
 ## 0.2.0
 
 * Restructure code